requirements.txt
pyptt
fastmcp
python-dotenv